backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
celery==5.3.6
redis==5.0.1
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
